#571 feat: Include humidity, motion and temperature as metrics
