+++ common/node/common-api/src/Metric.ts
@@ -1,0 +1,14 @@
+export enum MetricValue {
+    NONE = "none",
+    READ = "read",
+    VISIBLE = "visible",
+}
+
+export type Metric = {
+    current?: MetricValue;
+    humidity?: MetricValue;
+    motion?: MetricValue;
+    power?: MetricValue;
+    voltage?: MetricValue;
+    temperature?: MetricValue;
+};
